fix(ci): keep LSP runtime under LOC ratchet
